:racehorse: :art: Separated some routes and logic
