+++ app/Http/Controllers/Auth/AuthenticateController.php
@@ -97,32 +97,5 @@ class AuthenticateController extends Controller
     {
         //
     }
-    /**
-     * Get the current user (logged in) information
-     *
-     * @return void
-     */
-    public function userInfo()
-    {
-        $user = JWTAuth::parseToken()->authenticate();
 
-        // If the token is invalid
-        if (!$user) {
-            return response()->json([ 'invalid_user' ], 401);
-        }
-
-        return response()->json(
-            [
-              'first_name' =>   $user->first_name,
-              'last_name'  =>   $user->last_name,
-              'email'      =>   $user->email,
-              'username'   =>   $user->username,
-              'blood_type' =>   $user->donor->bloodType->code,
-              'avatar'     =>   '',//$user->avatar,
-              'birthdate'  =>   $user->birthdate,
-              'phone'      =>   $user->phone,
-              'bio'        =>   $user->bio
-            ], 200
-        );
-    }
 }
