+++ app/Models/User.php
@@ -111,4 +111,14 @@ class User extends Authenticatable
     {
         return $this->attributes['active'] == true;
     }
+
+    /**
+ * Get the route key for the model.
+ *
+ * @return string
+ */
+public function getRouteKeyName()
+{
+    return 'username';
+}
 }
