+++ database/seeds/CampaignsSeeder.php
@@ -1,6 +1,8 @@
 <?php
 
 use Illuminate\Database\Seeder;
+use DoeSangue\Models\User;
+use DoeSangue\Models\Campaign;
 
 class CampaignsSeeder extends Seeder
 {
@@ -13,15 +15,17 @@ class CampaignsSeeder extends Seeder
     {
         $this->command->info('Campanhas criadas!');
 
-        DB::table('campaigns')->truncate();
-        DB::table('campaigns')->insert(
-            [
-            array('title' => 'Campanha de teste do sistema', 'description' => 'Campanha teste! Campanha teste!Campanha teste!Campanha teste! Campanha teste! Campanha teste!Campanha teste!', 'image' => '', 'user_id' => '1', 'expires' => \Carbon\Carbon::now()->endOfYear(), 'created_at' => \Carbon\Carbon::now(), 'updated_at' => \Carbon\Carbon::now()),
-            array('title' => 'Solicitação de ajuda pela comunidade', 'description' => 'Campanha teste! Campanha teste! Campanha teste! Campanha teste!Campanha teste!Campanha teste!', 'image' => '', 'user_id' => '2', 'expires' => \Carbon\Carbon::now()->endOfYear(), 'created_at' => \Carbon\Carbon::now(), 'updated_at' => \Carbon\Carbon::now()),
-            array('title' => 'Solicitação de ajuda pela comunidade', 'description' => 'Campanha teste! Campanha teste! Campanha teste! Campanha teste!Campanha teste!Campanha teste!', 'image' => '', 'user_id' => '2', 'expires' => \Carbon\Carbon::now()->endOfYear(), 'created_at' => \Carbon\Carbon::now(), 'updated_at' => \Carbon\Carbon::now()),
-            array('title' => 'Solicitação de ajuda pela comunidade other', 'description' => 'Campanha teste! Campanha teste! Campanha teste! Campanha teste!Campanha teste!Campanha teste!', 'image' => '', 'user_id' => '2', 'expires' => \Carbon\Carbon::now()->endOfYear(), 'created_at' => \Carbon\Carbon::now(), 'updated_at' => \Carbon\Carbon::now()),
-            array('title' => 'Solicitação de ajuda pela comunidade one another', 'description' => 'Campanha teste! Campanha teste! Campanha teste! Campanha teste!Campanha teste!Campanha teste!', 'image' => '', 'user_id' => '2', 'expires' => \Carbon\Carbon::now()->endOfYear(), 'created_at' => \Carbon\Carbon::now(), 'updated_at' => \Carbon\Carbon::now()),
-            ]
-        );
+        DB::table('campaigns')->delete();
+
+          factory(User::class, 20)->create()->each(function ($u) {
+            $u->campaigns()->save(factory(Campaign::class)->make());
+          });
+            /*[
+              array('title' => 'Campanha de teste do sistema', 'description' => 'Campanha teste! Campanha teste!Campanha teste!Campanha teste! Campanha teste! Campanha teste!Campanha teste!', 'image' => '', 'user_id' => '1', 'expires' => \Carbon\Carbon::now()->endOfYear(), 'created_at' => \Carbon\Carbon::now(), 'updated_at' => \Carbon\Carbon::now()),
+              array('title' => 'Solicitação de ajuda pela comunidade', 'description' => 'Campanha teste! Campanha teste! Campanha teste! Campanha teste!Campanha teste!Campanha teste!', 'image' => '', 'user_id' => '2', 'expires' => \Carbon\Carbon::now()->endOfYear(), 'created_at' => \Carbon\Carbon::now(), 'updated_at' => \Carbon\Carbon::now()),
+              array('title' => 'Solicitação de ajuda pela comunidade', 'description' => 'Campanha teste! Campanha teste! Campanha teste! Campanha teste!Campanha teste!Campanha teste!', 'image' => '', 'user_id' => '2', 'expires' => \Carbon\Carbon::now()->endOfYear(), 'created_at' => \Carbon\Carbon::now(), 'updated_at' => \Carbon\Carbon::now()),
+              array('title' => 'Solicitação de ajuda pela comunidade other', 'description' => 'Campanha teste! Campanha teste! Campanha teste! Campanha teste!Campanha teste!Campanha teste!', 'image' => '', 'user_id' => '2', 'expires' => \Carbon\Carbon::now()->endOfYear(), 'created_at' => \Carbon\Carbon::now(), 'updated_at' => \Carbon\Carbon::now()),
+              array('title' => 'Solicitação de ajuda pela comunidade one another', 'description' => 'Campanha teste! Campanha teste! Campanha teste! Campanha teste!Campanha teste!Campanha teste!', 'image' => '', 'user_id' => '2', 'expires' => \Carbon\Carbon::now()->endOfYear(), 'created_at' => \Carbon\Carbon::now(), 'updated_at' => \Carbon\Carbon::now()),
+            ]*/
     }
 }
